make chart a seperate file
